Sorting fix

+++ ExtractIndirectCoupling/ProjectParser/JsonChainStream.cs
@@ -12,9 +12,9 @@ namespace ProjectParser
         private string streamName;
         private BufferedStream chainStream;
         private JsonChainStream sourceStream;
-        private int position;
+        private long position;
         private int count;
-        private int sourceStart;
+        private long sourceStart;
         private int offset;
         private bool merged;
         private int level;
@@ -25,7 +25,7 @@ namespace ProjectParser
 
         public string StreamName { get => streamName; set => streamName = value; }
         public BufferedStream ChainStream { get => chainStream; set => chainStream = value; }
-        public int Position { get => position; set => position = value; }
+        public long Position { get => position; set => position = value; }
         public int Count { get => count; set => count = value; }
         public int Offset { get => offset; set => offset = value; }
         public bool Merged { get => merged; set => merged = value; }
@@ -33,11 +33,11 @@ namespace ProjectParser
         public int Index { get => index; set => index = value; }
         public int Threshold { get => threshold; set => threshold = value; }
         public JsonChainStream SourceStream { get => sourceStream; set => sourceStream = value; }
-        public int SourceStart { get => sourceStart; set => sourceStart = value; }
+        public long SourceStart { get => sourceStart; set => sourceStart = value; }
         internal JsonSubchain NextChain { get => nextChain; set => nextChain = value; }
         internal IComparer<JsonSubchain> Comparer { get => comparer; set => comparer = value; }
 
-        public void Open(string filename, IComparer<JsonSubchain> comparer)
+        public void Open(string filename, IComparer<JsonSubchain> comparer, bool delete = false)
         {
             StreamName = filename;
             Position = 0;
@@ -47,13 +47,24 @@ namespace ProjectParser
             Merged = false;
             Level = 0;
             Index = 0;
-            ChainStream = new BufferedStream(File.Open(StreamName, FileMode.Create), 4096 * Offset);
+            if (delete && File.Exists(StreamName))
+            {
+               try
+                {
+                    File.Delete(StreamName);
+                }
+                catch (Exception e)
+                {
+                    // ignore
+                }
+            }
+            ChainStream = new BufferedStream(File.Open(StreamName, FileMode.OpenOrCreate, FileAccess.ReadWrite, FileShare.ReadWrite), 512000 * Offset);
             SourceStream = this;
-            Threshold = 100000;
+            Threshold = 1024000000 / JsonSubchainRecord.ByteArrayLen;
             Comparer = comparer;
         }
 
-        public void Open(JsonChainStream stream, int sourceStart, int count, int level)
+        public void Open(JsonChainStream stream, long sourceStart, int count, int level, bool delete = false)
         {
             StreamName = stream.StreamName + (++stream.Index).ToString();
             Position = 0;
@@ -63,13 +74,24 @@ namespace ProjectParser
             Merged = true;
             Level = level + 1;
             Index = stream.Index;
-            ChainStream = new BufferedStream(File.Open(StreamName, FileMode.Create));
+            if (delete && File.Exists(StreamName))
+            {
+                try
+                {
+                    File.Delete(StreamName);
+                }
+                catch (Exception e)
+                {
+                    // ignore
+                }
+            }
+            ChainStream = new BufferedStream(File.Open(StreamName, FileMode.OpenOrCreate, FileAccess.ReadWrite, FileShare.ReadWrite));
             SourceStream = stream.SourceStream;
             Threshold = stream.Threshold;
             Comparer = stream.Comparer;
         }
 
-        public void Close()
+        public void Close(bool delete = false)
         {
             if (ChainStream != null)
             {
@@ -77,13 +99,26 @@ namespace ProjectParser
                 ChainStream = null;
             }
 
-            if (File.Exists(StreamName))
-                File.Delete(StreamName);
+            try
+            {
+                if (delete && File.Exists(StreamName))
+                    File.Delete(StreamName);
+            }
+            catch (Exception e)
+            {
+                // ignore
+            }
         }
 
-        public void Seek(int position)
+        public void Seek(long position)
         {
             ChainStream.Seek(position * Offset, SeekOrigin.Begin);
+            Position = position;
+        }
+
+        public void Flush()
+        {
+            ChainStream.Flush();
         }
 
         public bool HasNext()
@@ -99,11 +134,11 @@ namespace ProjectParser
         public void Write(JsonSubchain chain)
         {
             ChainStream.Write(JsonSubchainRecord.ObjectToByteArray(new JsonSubchainRecord(chain)), 0, Offset);
+            if (Position >= Count) Count++;
             Position++;
-            Count++;
         }
 
-        public void Write(JsonSubchain chain, int position)
+        public void Write(JsonSubchain chain, long position)
         {
             ChainStream.Seek(position * Offset, SeekOrigin.Begin);
             ChainStream.Write(JsonSubchainRecord.ObjectToByteArray(new JsonSubchainRecord(chain)), 0, Offset);
@@ -113,12 +148,13 @@ namespace ProjectParser
         public JsonSubchain Read()
         {
             byte[] bytes = new byte[Offset];
+            //ChainStream.Seek(Position * Offset, SeekOrigin.Begin);
             ChainStream.Read(bytes, 0, Offset);
             Position++;
             return JsonSubchainRecord.ByteArrayToObject(bytes).GetJsonSubchain();
         }
 
-        public JsonSubchain Read(int position)
+        public JsonSubchain Read(long position)
         {
             byte[] bytes = new byte[Offset];
             ChainStream.Seek(position * Offset, SeekOrigin.Begin);
@@ -127,20 +163,83 @@ namespace ProjectParser
             return JsonSubchainRecord.ByteArrayToObject(bytes).GetJsonSubchain();
         }
 
-        public JsonSubchain[] Read(int position, int count)
+        public JsonSubchain[] Read(long position, long count)
         {
             byte[] bytes = new byte[Offset];
             byte[] buffer = new byte[Offset * count];
             JsonSubchain[] chains = new JsonSubchain[count];
             ChainStream.Seek(position * Offset, SeekOrigin.Begin);
-            ChainStream.Read(buffer, 0, Offset * count);
+            ChainStream.Read(buffer, 0, Offset * (int)count);
             Position = position + count;
-            for (int i = 0; i < count; i++)
+            for (long i = 0; i < count; i++)
             {
                 Array.Copy(buffer, i * Offset, bytes, 0, Offset);
                 chains[i] = JsonSubchainRecord.ByteArrayToObject(bytes).GetJsonSubchain();
             }
             return chains;
+        }
+
+        public void CheckSort()
+        {
+            Seek(0);
+            long minK = (Count - 1) / Threshold;
+            long maxK = (Count + Threshold - 1) / Threshold;
+            bool init = true;
+            int next;
+            JsonSubchain subchain = null;
+            JsonSubchain[] chains = new JsonSubchain[Threshold];
+
+            for (int i = 0; i < minK; i++)
+            {
+                chains = Read(Position, Threshold);
+                next = 0;
+                if (init)
+                {
+                    init = false;
+                    subchain = chains[0];
+                    next++;
+                }
+                for (int j = next; j < Threshold; j++)
+                {
+                    if (Comparer.Compare(subchain, chains[j]) > 0)
+                    {
+                        bool stopHere = true;
+                        Console.Write("****");
+                    }
+                    Console.WriteLine("from: {0} to: {1} initial: {2} final: {3}", 
+                        subchain.From.ToString().PadLeft(4),
+                        subchain.To.ToString().PadLeft(4),
+                        subchain.Initial.ToString().PadRight(5), 
+                        subchain.Final.ToString().PadRight(5));
+                    subchain = chains[j];
+                }
+            }
+
+            long q = Threshold - ((maxK * Threshold) - Count);
+            if (q > 0)
+            {
+                chains = Read(Position, q);
+                next = 0;
+                if (init)
+                {
+                    subchain = chains[0];
+                    next++;
+                }
+                for (int j = next; j < q; j++)
+                {
+                    if (Comparer.Compare(subchain, chains[j]) > 0)
+                    {
+                        bool stopHere = true;
+                        Console.Write("****");
+                    }
+                    Console.WriteLine("from: {0} to: {1} initial: {2} final: {3}",
+                        subchain.From.ToString().PadLeft(4),
+                        subchain.To.ToString().PadLeft(4),
+                        subchain.Initial.ToString().PadRight(5),
+                        subchain.Final.ToString().PadRight(5));
+                    subchain = chains[j];
+                }
+            }
         }
 
         public JsonChainStream MergeSort()
@@ -161,13 +260,13 @@ namespace ProjectParser
 
                 Merge(this, left, right);
 
-                left.Close();
-                right.Close();
+                left.Close(true);
+                right.Close(true);
             }
             return this;
         }
 
-        public void Sort()
+        private void Sort()
         {
             JsonSubchain[] subchains = SourceStream.Read(SourceStart, Count);
             Array.Sort(subchains, Comparer);
@@ -178,7 +277,7 @@ namespace ProjectParser
             }
         }
 
-        public void Merge(JsonChainStream result, JsonChainStream left, JsonChainStream right)
+        private void Merge(JsonChainStream result, JsonChainStream left, JsonChainStream right)
         {
             left.Seek(0);
             right.Seek(0);
@@ -191,7 +290,7 @@ namespace ProjectParser
 
                 while (true)
                 {
-                    if (Comparer.Compare(left.NextChain, right.NextChain) > -1)
+                    if (Comparer.Compare(left.NextChain, right.NextChain) < 1)
                     {
                         result.Write(left.NextChain);
                         if (left.HasNext())
